Removed a few redundant casts from the ExtractXMP method.

+++ HostTest/MetaDataHelper.cs
@@ -863,7 +863,7 @@ namespace HostTest
 					if (signature == TIFFSignature)
 					{
 						uint ifdOffset = ReadLong(stream, littleEndian);
-						stream.Seek((long)ifdOffset, SeekOrigin.Begin);
+						stream.Seek(ifdOffset, SeekOrigin.Begin);
 
 						int ifdCount = ReadShort(stream, littleEndian);
 
@@ -873,7 +873,7 @@ namespace HostTest
 
 							if (ifd.tag == XmpTag && (ifd.type == DataType.Byte || ifd.type == DataType.Undefined))
 							{
-								stream.Seek((long)ifd.offset, SeekOrigin.Begin);
+								stream.Seek(ifd.offset, SeekOrigin.Begin);
 
 								int count = (int)ifd.count;
